edited customer rep

+++ Amarath/DAL/Data/CustomerRepository.cs
@@ -17,13 +17,18 @@ namespace Amarath.DAL.Data
         public Customer Add(Customer customer)
         {
             context.Customers.Add(customer);
+            context.SaveChanges();
             return customer;
         }
 
         public Customer Delete(int CustomerId)
         {
             Customer customer = context.Customers.Find(CustomerId);
-            context.Customers.Remove(customer);
+            if(customer != null)
+            {
+                context.Customers.Remove(customer);
+                context.SaveChanges();
+            }
             return customer;
         }
 
@@ -37,10 +42,12 @@ namespace Amarath.DAL.Data
             return context.Customers.Find(CustomerId);
         }
 
-        public Customer Update(Customer customer)
+        public Customer Update(Customer customerChanges)
         {
-            context.Entry(customer).State = Microsoft.EntityFrameworkCore.EntityState.Modified;
-            return customer;
+            var customer = context.Customers.Attach(customerChanges);
+            customer.State = Microsoft.EntityFrameworkCore.EntityState.Modified;
+            context.SaveChanges();
+            return customerChanges;
         }
         public void Save()
         {
